Hooks simplified, Ext Authorizer teardown
